+++ neusler/search/api/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .views import SearchAPI
+
+urlpatterns = [
+    path("", SearchAPI.as_view(), name="search"),
+]
