+++ neusler/search/views.py
@@ -1,14 +1,10 @@
 from django.core.paginator import EmptyPage, PageNotAnInteger, Paginator
 from django.template.response import TemplateResponse
 
-from rest_framework.response import Response
-from rest_framework.views import APIView
 from wagtail.core.models import Page
 from wagtail.search.models import Query
 
-from neusler.cms.models import ArticlePage, PostPage, VideoPage
-
-from .serializers import ArticleResultsSerializer, VideoResultsSerializer
+from neusler.cms.models import PostPage
 
 
 def search(request):
@@ -41,22 +37,3 @@ def search(request):
             "search_results": search_results,
         },
     )
-
-
-class SearchAPI(APIView):
-    def get(self, request, *args, **kwargs):
-        search_query = request.data.get("query", None)
-        if search_query:
-            article_results = ArticlePage.objects.live().search(search_query)
-            video_results = VideoPage.objects.live().search(search_query)
-
-            query = Query.get(search_query)
-
-            # Record hit
-            query.add_hit()
-        else:
-            article_results = Page.objects.none()
-            video_results = Page.objects.none()
-        article_serializer = ArticleResultsSerializer(article_results, many=True)
-        video_serializer = VideoResultsSerializer(video_results, many=True)
-        return Response([video_serializer.data, article_serializer.data])
